+++ src/game/nlcore/elements/transition/type.ts
@@ -10,6 +10,7 @@ export type ElementProp<T extends Element = Element, U extends React.HTMLAttribu
     & U;
 export type ImgElementProp = ElementProp<HTMLImageElement, React.ImgHTMLAttributes<HTMLImageElement>>;
 export type SpanElementProp = ElementProp<HTMLSpanElement, React.HTMLAttributes<HTMLSpanElement>>;
+export type DivElementProp = ElementProp<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>;
 export type CSSElementProp<T extends React.CSSProperties | DOMKeyframesDefinition> = ElementProp & { style: T };
 export type CSSProps = React.CSSProperties;